requirements.txt
langchain_openai==1.1.12
langchain_core==1.2.28
python-dotenv==1.2.2
streamlit==1.56.0
langchain_community==0.4.1